requirements.txt
autopep8==1.3.4
configparser==4.0.2
matplotlib==2.1.1
pep8==1.7.1
psycopg2-binary==2.8.4
pycrypto==2.6.1
pylint==1.9.5
simple-crypt==4.1.7

